coreService/requirements.txt
pymongo==3.10.1
python-dotenv==0.11.0
flask-restx==0.1.1
Flask==1.1
regex==2020.2.20